requirements.txt
fastapi
uvicorn[standard]
aiofiles
jinja2
line-bot-sdk
responses
pytest
pytest-flake8
pytest-pep8
pytest-cov
mock
ffmpeg-python
google-cloud-speech
google-cloud-storage